% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbClearResult,GDALOGRSQLResult-method}
\alias{dbClearResult,GDALOGRSQLResult-method}
\title{Clear Results}
\usage{
\S4method{dbClearResult}{GDALOGRSQLResult}(res, ...)
}
\arguments{
\item{res}{A GDALOGRSQLResult Object}

\item{...}{Additional arguments}
}
\description{
Clear Results
}
