% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbColumnInfo,GDALOGRSQLResult-method}
\alias{dbColumnInfo,GDALOGRSQLResult-method}
\title{Get names and types of the result set’s columns.}
\usage{
\S4method{dbColumnInfo}{GDALOGRSQLResult}(res)
}
\arguments{
\item{res}{GDALOGRSQLResult}
}
\description{
Get names and types of the result set’s columns.
}
