% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbDataType,GDALOGRSQLConnection-method}
\alias{dbDataType,GDALOGRSQLConnection-method}
\title{Find the database data type associated with an R object}
\usage{
\S4method{dbDataType}{GDALOGRSQLConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{GDALOGRConnection}

\item{obj}{An R object}

\item{...}{Additional arguments}
}
\description{
Find the database data type associated with an R object
}
