% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbDisconnect,GDALOGRSQLConnection-method}
\alias{dbDisconnect,GDALOGRSQLConnection-method}
\title{Disconnect a GDALOGRSQLConnection}
\usage{
\S4method{dbDisconnect}{GDALOGRSQLConnection}(conn, ...)
}
\arguments{
\item{conn}{A GDALOGRSQLConnection}

\item{...}{Additional arguments (not used)}
}
\description{
Disconnect a GDALOGRSQLConnection
}
