% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbFetch,GDALOGRSQLResult-method}
\alias{dbFetch,GDALOGRSQLResult-method}
\title{Retrieve records from OGRSQL query}
\usage{
\S4method{dbFetch}{GDALOGRSQLResult}(res, n = NULL, ..., geom = "wkt", fid = FALSE)
}
\arguments{
\item{res}{A GDALOGRSQLResult Object}

\item{n}{Number of records to fetch. Default \code{-1} for all records.}

\item{...}{Additional arguments}

\item{geom}{character. Either \code{"json"}, \code{"gml"}, \code{"kml"}, \code{"wkt"} (default) or \code{"hex"}}

\item{fid}{logical. Keep feature ID? Default: \code{FALSE}}
}
\description{
Retrieve records from OGRSQL query
}
