% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbGetInfo,GDALOGRSQLConnection-method}
\alias{dbGetInfo,GDALOGRSQLConnection-method}
\title{Get GDALOGRSQLConnection information}
\usage{
\S4method{dbGetInfo}{GDALOGRSQLConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{A GDALOGRSQLConnection}

\item{...}{Additional arguments (not used)}
}
\description{
Get GDALOGRSQLConnection information
}
