% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbGetInfo,GDALOGRSQLDriver-method}
\alias{dbGetInfo,GDALOGRSQLDriver-method}
\title{Get GDALOGRSQLDriver driver and GDAL version information}
\usage{
\S4method{dbGetInfo}{GDALOGRSQLDriver}(dbObj, ...)
}
\arguments{
\item{dbObj}{A GDALOGRSQLDriver}

\item{...}{Additional arguments (not used)}
}
\description{
Get GDALOGRSQLDriver driver and GDAL version information
}
