% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbGetRowCount,GDALOGRSQLResult-method}
\alias{dbGetRowCount,GDALOGRSQLResult-method}
\title{Get number of rows returned in a SELECT query}
\usage{
\S4method{dbGetRowCount}{GDALOGRSQLResult}(res)
}
\arguments{
\item{res}{GDALOGRSQLResult}
}
\description{
Get number of rows returned in a SELECT query
}
