% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbGetRowsAffected,GDALOGRSQLResult-method}
\alias{dbGetRowsAffected,GDALOGRSQLResult-method}
\title{Get number of rows altered in a INSERT/UPDATE query}
\usage{
\S4method{dbGetRowsAffected}{GDALOGRSQLResult}(res)
}
\arguments{
\item{res}{GDALOGRSQLResult}
}
\description{
Get number of rows altered in a INSERT/UPDATE query
}
