% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbGetStatement,GDALOGRSQLResult-method}
\alias{dbGetStatement,GDALOGRSQLResult-method}
\title{Get the issued query as a character value}
\usage{
\S4method{dbGetStatement}{GDALOGRSQLResult}(res)
}
\arguments{
\item{res}{GDALOGRSQLResult}
}
\description{
Get the issued query as a character value
}
