% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbHasCompleted,GDALOGRSQLResult-method}
\alias{dbHasCompleted,GDALOGRSQLResult-method}
\title{Check if rows remain to be fetched}
\usage{
\S4method{dbHasCompleted}{GDALOGRSQLResult}(res, ...)
}
\arguments{
\item{res}{A GDALOGRSQLResult Object}

\item{...}{Additional arguments}
}
\description{
Check if rows remain to be fetched
}
