% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbIsValid,GDALOGRSQLConnection-method}
\alias{dbIsValid,GDALOGRSQLConnection-method}
\title{Determine if connection is open or closed}
\usage{
\S4method{dbIsValid}{GDALOGRSQLConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{GDALOGRConnection}

\item{...}{Additional arguments}
}
\description{
Determine if connection is open or closed
}
