% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbIsValid,GDALOGRSQLResult-method}
\alias{dbIsValid,GDALOGRSQLResult-method}
\title{Determine if a result set is open or closed}
\usage{
\S4method{dbIsValid}{GDALOGRSQLResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{GDALOGRSQLResult}

\item{...}{Additional arguments}
}
\description{
Determine if a result set is open or closed
}
