% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbListFields,GDALOGRSQLConnection,character-method}
\alias{dbListFields,GDALOGRSQLConnection,character-method}
\title{Get fields are available in a given table.}
\usage{
\S4method{dbListFields}{GDALOGRSQLConnection,character}(conn, name)
}
\arguments{
\item{conn}{GDALOGRSQLConnection}

\item{name}{character. Table name.}
}
\description{
Get fields are available in a given table.
}
