% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{dbSendQuery,GDALOGRSQLConnection,ANY-method}
\alias{dbSendQuery,GDALOGRSQLConnection,ANY-method}
\title{Send a query to GDALOGRSQLDriver}
\usage{
\S4method{dbSendQuery}{GDALOGRSQLConnection,ANY}(conn, statement, ...)
}
\arguments{
\item{conn}{A GDALOGRSQLConnection Object}

\item{statement}{An OGRSQL query to execute.}

\item{...}{Additional arguments}
}
\description{
Send a query to GDALOGRSQLDriver
}
\examples{
# This is another good place to put examples
}
