% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\docType{class}
\name{GDALOGRSQLConnection-class}
\alias{GDALOGRSQLConnection-class}
\alias{dbQuoteIdentifier,GDALOGRSQLConnection,character-method}
\title{DBI Connection for GDAL OGRSQL}
\usage{
\S4method{dbQuoteIdentifier}{GDALOGRSQLConnection,character}(conn, x, ...)
}
\description{
DBI Connection for GDAL OGRSQL

Safely quote strings and identifiers to avoid SQL injection attacks
}
\keyword{internal}
