% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\docType{class}
\name{GDALOGRSQLDriver-class}
\alias{GDALOGRSQLDriver-class}
\alias{dbUnloadDriver,GDALOGRSQLDriver-method}
\title{DBI Driver for GDAL OGRSQL}
\usage{
\S4method{dbUnloadDriver}{GDALOGRSQLDriver}(drv, ...)
}
\description{
DBI Driver for GDAL OGRSQL
}
\keyword{internal}
