% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{OGRSQL}
\alias{OGRSQL}
\alias{dbConnect,GDALOGRSQLDriver-method}
\title{GDAL OGR SQL Driver}
\usage{
OGRSQL()

\S4method{dbConnect}{GDALOGRSQLDriver}(drv, dsn, bigint = "integer64", ...)
}
\arguments{
\item{drv}{An object created by \code{OGRSQL())}}

\item{dsn}{character. Path to OGR-compatible data source.}

\item{bigint}{character. Return type for \emph{OFTInteger64} fields. One of: \code{"integer"}, \code{"character"}, \code{"numeric"}, \code{"integer64"} (default)}

\item{...}{Additional arguments}
}
\description{
GDAL OGR SQL Driver
}
\examples{
\dontrun{
db <- dbConnect(ROGRSQL::OGRSQL(), "nc.gpkg")
dbSendQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
