% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROGRSQL.R
\name{sample_gpkg_path}
\alias{sample_gpkg_path}
\title{Sample 'Luxembourg' Geopackage}
\usage{
sample_gpkg_path()
}
\value{
character. Path to ROGRSQL package \code{"extdata"} folder \code{"lux.gpkg"} file.
}
\description{
This is a Geopackage file created from the \code{"lux.shp"} test dataset in the \code{{terra}} R package.
}
\examples{
sample_gpkg_path()
}
\references{
Hijmans R (2024). \emph{terra: Spatial Data Analysis}. R package version 1.7-78, https://rspatial.github.io/terra/, \url{https://rspatial.org/}.
}
